Refined the frontend by implementing dashboard UI components
