+++ frontend/components/admin/settings/PaymentSettings.tsx
@@ -5,7 +5,7 @@ import { Input } from "@/components/ui/input"
 import { Label } from "@/components/ui/label"
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select"
 import { CreditCard } from "lucide-react"
-import { SystemSettings } from "@/types/settings"
+import { SystemSettings } from "@/lib/api"
 
 interface PaymentSettingsProps {
     settings: SystemSettings
