+++ frontend/components/admin/settings/SystemControl.tsx
@@ -4,7 +4,7 @@ import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card"
 import { Label } from "@/components/ui/label"
 import { Button } from "@/components/ui/button"
 import { Settings } from "lucide-react"
-import { SystemSettings } from "@/types/settings"
+import { SystemSettings } from "@/lib/api"
 
 interface SystemControlProps {
     settings: SystemSettings
