+++ frontend/lib/constants/labels.ts
@@ -1,0 +1,19 @@
+export const LABELS = {
+  STATUS: {
+    ACTIVE: "Active",
+    EXPIRED: "Expired",
+    BLOCKED: "Blocked",
+    UNUSED: "Unused",
+    FULLY_USED: "Fully Used",
+    COMPLETED: "Completed",
+    PENDING: "Pending",
+    FAILED: "Failed",
+    REFUNDED: "Refunded",
+  },
+  PLANS: [
+    { key: "1Hr",   label: "1 Hour",   price: "KSh 10" },
+    { key: "4Hrs",  label: "4 Hours",  price: "KSh 15" },
+    { key: "12Hrs", label: "12 Hours", price: "KSh 20" },
+    { key: "24Hrs", label: "24 Hours", price: "KSh 30" },
+  ]
+}
